ai_cam_client/requirements.txt
numpy>=2.1.0
opencv-python>=4.10.0.84
ultralytics>=8.3.0
Pillow>=10.2.0
einops
torch
transformers
accelerate
